refactor: domain 분리
